ok, added `inavid_operator_args` and associated stuff
