+++ Common/src/main/resources/assets/hexcasting/lang/en_us.flatten.json5
@@ -890,6 +890,11 @@
         cos: "the cosine of %s",
       },
 
+      invalid_operator_args: {
+        single: "Got an unexpected iota at index %d of the stack: %s",
+        plural: "Got %d unexpected iotas at indices %d-%d of the stack: %s",
+      },
+
       no_akashic_record: "No Akashic Record at %s",
       disallowed: "has been disallowed by the server admins",
       disallowed_circle: "has been disallowed in spell circles by the server admins",
